Break ingredients into a new component
