+++ src/components/Recipe/Recipe.css
@@ -48,55 +48,3 @@
 .recipe__about__description {
   padding: 10px;
 }
-
-.recipe__ingredients {
-  padding: 10px 5px;
-  background-color: pink;
-}
-
-.malt__row {
-  border-bottom: 1px solid black;
-  text-align: center;
-}
-
-.ingredient-name {
-  text-align: left;
-}
-
-.malt-table,
-.hop-table,
-.yeast-table {
-  border-collapse: collapse;
-  width: 100%;
-  margin-bottom: 30px;
-}
-
-.table-head {
-  display: flex;
-  align-items: center;
-  font-size: 18px;
-  padding-bottom: 5px;
-}
-
-.icon {
-  display: inline-block;
-  display: flex;
-  justify-content: center;
-  align-items: center;
-  width: 50px;
-  height: 50px;
-  background-color: black;
-  border-radius: 100%;
-  margin-right: 10px;
-}
-
-.icon__malt,
-.icon__hop,
-.icon__yeast {
-  fill: white;
-  font-size: 30px;
-}
-
-tbody {
-  line-height: 1.8;
-}
